backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
joblib==1.3.2
scikit-learn==1.3.2
xgboost==2.0.3
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
pandas>=1.0.0
numpy<2.0
matplotlib>=3.0.0
seaborn>=0.11.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt==4.1.1
email-validator>=2.0.0
